backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
psycopg2-binary
